LayoutTests/fast/events/onunload-not-on-body-expected.txt

ALERT: unload
you should only see one unload alert appear.
ALERT: unload
you should only see one unload alert appear.
